+++ CHANGELOG.md
@@ -6,6 +6,7 @@ Changelog for NeoFS Node
 ### Added
 
 ### Fixed
+- Bearer token signed not by its issuer is no longer passed (#3216)
 
 ### Changed
 - IR calls `ObjectService.SearchV2` to select SG objects now (#3144)